add styled components and refactoring speakers section
