+++ package.json
@@ -8,13 +8,15 @@
     "gh-pages": "^1.0.0",
     "glamor": "^2.20.30",
     "react": "^16.8.6",
-    "react-dom": "^16.8.6",
     "react-addons-test-utils": "^15.6.0",
+    "react-dom": "^16.8.6",
     "react-html5video": "^2.3.0",
+    "react-modal-video": "^1.2.2",
     "react-responsive": "^1.3.4",
     "react-snapshot": "^1.0.4",
     "react-svg-morph": "^0.1.10",
-    "react-video-cover": "^1.1.0"
+    "react-video-cover": "^1.1.0",
+    "styled-components": "^4.2.0"
   },
   "devDependencies": {
     "husky": "^0.14.3",
