+++ src/utils/Globals.js
@@ -21,8 +21,8 @@ const Globals = {
     midMain: '#003F5F',
     lightMain: '#BFCAFE',
     primary: '#220132',
+    secondary: '#FA46FF',
     header: '#222',
-    secondary: '#555555',
     subheader: '#555555',
     topic: '#222',
     darken: '#222',
